user_service/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
pymongo
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
python-dotenv
python-multipart